# apps/frontend/app/utils/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "utoipa-axum",
    "description": "Utoipa's axum bindings for seamless integration for the two",
    "contact": {
      "name": "Juha Kukkonen",
      "email": "juha7kukkonen@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/tenants/{tenant_id}/projects": {
      "get": {
        "summary": "プロジェクト一覧",
        "operationId": "list_projects",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "テナントID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "プロジェクト一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/crate.entities.projects.Model"
                  }
                }
              }
            }
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "プロジェクトを作成",
        "operationId": "create_project",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "テナントID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "作成されたプロジェクト",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.entities.projects.Model"
                }
              }
            }
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/tenants/{tenant_id}/projects/{id}": {
      "get": {
        "summary": "プロジェクトを取得",
        "operationId": "get_project",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "テナントID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "プロジェクトID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "プロジェクト情報",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.entities.projects.Model"
                }
              }
            }
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "プロジェクトを更新",
        "operationId": "update_project",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "テナントID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "プロジェクトID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "更新後のプロジェクト",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/crate.entities.projects.Model"
                }
              }
            }
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "プロジェクトを削除",
        "operationId": "delete_project",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "テナントID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "プロジェクトID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "削除しました"
          },
          "401": {
            "description": "ログインまたはセッションが必要です",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "この操作は許可されていません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "リソースが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "サーバー側で問題が発生しました。時間をおいて再度お試しください",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "API 共通のエラー応答ボディ。",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "internal-error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProjectRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "icon_emoji": {
            "type": [
              "string",
              "null"
            ]
          },
          "icon_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UpdateProjectRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "icon_emoji": {
            "type": [
              "string",
              "null"
            ]
          },
          "icon_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "crate.entities.projects.Model": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "tenant_id"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "icon_emoji": {
            "type": [
              "string",
              "null"
            ]
          },
          "icon_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  }
}